Move nav destination logic to feature modules impl
This moves most of the navigation related code away from the app module into feature modules which makes things a bit cleaner
