+++ apps/browser/build.gradle
@@ -67,9 +67,6 @@ dependencies {
     implementation libs.navigation.fragment.ktx
     implementation libs.navigation.ui.ktx
 
-    // Serialization
-    implementation libs.gson
-
     // Features
     implementation project(":features:projects:impl")
