+++ features/projects/impl/build.gradle
@@ -40,6 +40,7 @@ dependencies {
     api project(":features:users:api")
     implementation project(":shared:service:api")
     implementation project(":shared:navigation:api")
+    implementation project(":shared:navigation:impl")
 
     // Recyclerview convenience
     implementation libs.groupie.viewbinding
@@ -59,10 +60,6 @@ dependencies {
     ksp libs.dagger.hilt.compiler
     implementation libs.dagger.android
     ksp libs.dagger.android.compiler
-
-    // Navigation
-    implementation libs.navigation.fragment.ktx
-    implementation libs.navigation.ui.ktx
 
     // Serialization
     implementation libs.gson
